+++ c-fractals/src/gradients.h
@@ -11,22 +11,22 @@
 
 
 /* convert a fractal value to rgb - ultra */
-void fractal_value_to_rgb_ultra(float * r, float * g, float * b, int value);
+void fractal_value_to_rgb_ultra(uint8_t * r, uint8_t * g, uint8_t * b, int value);
 
 /* convert a fractal value to rgb - monochrome */
-void fractal_value_to_rgb_monochrome(float * r, float * g, float * b, int value);
+void fractal_value_to_rgb_monochrome(uint8_t * r, uint8_t * g, uint8_t * b, int value);
 
 /* convert a fractal value to rgb - tri */
-void fractal_value_to_rgb_tri(float * r, float * g, float * b, int value);
+void fractal_value_to_rgb_tri(uint8_t * r, uint8_t * g, uint8_t * b, int value);
 
 /* convert a fractal value to rgb - jet */
-void fractal_value_to_rgb_jet(float * r, float * g, float * b, int value);
+void fractal_value_to_rgb_jet(uint8_t * r, uint8_t * g, uint8_t * b, int value);
 
 /* convert a fractal value to rgb - lavender */
-void fractal_value_to_rgb_lavender(float * r, float * g, float * b, int value);
+void fractal_value_to_rgb_lavender(uint8_t * r, uint8_t * g, uint8_t * b, int value);
 
 /* convert a fractal value to rgb - binary */
-void fractal_value_to_rgb_binary(float * r, float * g, float * b, int value);
+void fractal_value_to_rgb_binary(uint8_t * r, uint8_t * g, uint8_t * b, int value);
 
 
 #endif
